+++ gcdt/kumo_core.py
@@ -1,22 +1,24 @@
 # -*- coding: utf-8 -*-
 from __future__ import print_function
-import os
-import sys
-import random
-import string
-import json
-import time
-import six
+
 import imp
+import json
+import os
+import random
+import six
+import string
+import sys
+import time
 from datetime import tzinfo, timedelta, datetime
+
 import pyhocon.exceptions
-from pyhocon.exceptions import ConfigMissingException
-from pyhocon import ConfigFactory
-from tabulate import tabulate
-from pyspin.spin import Default, Spinner
 from clint.textui import colored, prompt
-from glomex_utils.config_reader import get_env
 from gcdt import monitoring
+from glomex_utils.config_reader import get_env
+from pyhocon import ConfigFactory
+from pyhocon.exceptions import ConfigMissingException
+from pyspin.spin import Default, Spinner
+from tabulate import tabulate
 
 
 def load_cloudformation_template(path=None):
@@ -304,21 +306,23 @@ def _stack_exists(boto_session, stackName):
         return True
 
 
-def deploy_stack(boto_session, conf, cloudformation, slack_token,
-                 override_stack_policy=False):
+def deploy_stack(boto_session, conf, cloudformation, slack_token=None,
+                 slack_channel='systemmessages', override_stack_policy=False):
     """Deploy the stack to AWS cloud. Does either create or update the stack.
 
     :param conf:
     :param slack_token:
+    :param slack_channel:
     :param override_stack_policy:
     :return: exit_code
     """
     stackname = _get_stack_name(conf)
     if _stack_exists(boto_session, stackname):
         return _update_stack(boto_session, conf, cloudformation,
-                             override_stack_policy, slack_token)
+                             override_stack_policy, slack_token, slack_channel)
     else:
-        return _create_stack(boto_session, conf, cloudformation, slack_token)
+        return _create_stack(boto_session, conf, cloudformation, slack_token,
+                             slack_channel)
 
 
 def _s3_upload(boto_session, conf, cloudformation):
@@ -408,7 +412,8 @@ def _get_stack_policy_during_update(cloudformation, override_stack_policy):
     return stack_policy_during_update
 
 
-def _create_stack(boto_session, conf, cloudformation, slack_token):
+def _create_stack(boto_session, conf, cloudformation, slack_token=None,
+                  slack_channel='systemmessages'):
     # create stack with all the information we have
     client_cf = boto_session.client('cloudformation')
     _call_pre_create_hook(cloudformation)
@@ -436,7 +441,7 @@ def _create_stack(boto_session, conf, cloudformation, slack_token):
         )
 
     message = 'kumo bot: created stack %s ' % _get_stack_name(conf)
-    monitoring.slacker_notification('systemmessages', message, slack_token)
+    monitoring.slack_notification(slack_channel, message, slack_token)
     stackname = _get_stack_name(conf)
     exit_code = _poll_stack_events(boto_session, stackname)
     _call_post_create_hook(cloudformation)
@@ -444,7 +449,8 @@ def _create_stack(boto_session, conf, cloudformation, slack_token):
     return exit_code
 
 
-def _update_stack(boto_session, conf, cloudformation, override_stack_policy, slack_token):
+def _update_stack(boto_session, conf, cloudformation, override_stack_policy,
+                  slack_token=None, slack_channel='systemmessages'):
     # update stack with all the information we have
     exit_code = 0
     client_cf = boto_session.client('cloudformation')
@@ -480,7 +486,7 @@ def _update_stack(boto_session, conf, cloudformation, override_stack_policy, sla
             )
 
         message = 'kumo bot: updated stack %s ' % _get_stack_name(conf)
-        monitoring.slacker_notification('systemmessages', message, slack_token)
+        monitoring.slack_notification(slack_channel, message, slack_token)
         stackname = _get_stack_name(conf)
         exit_code = _poll_stack_events(boto_session, stackname)
         _call_post_update_hook(cloudformation)
@@ -495,19 +501,21 @@ def _update_stack(boto_session, conf, cloudformation, override_stack_policy, sla
     return exit_code
 
 
-def delete_stack(boto_session, conf, slack_token):
+def delete_stack(boto_session, conf, slack_token=None,
+                 slack_channel='systemmessages'):
     """Delete the stack from AWS cloud.
 
     :param boto_session:
     :param conf:
     :param slack_token:
+    :param slack_channel:
     """
     client_cf = boto_session.client('cloudformation')
     response = client_cf.delete_stack(
         StackName=_get_stack_name(conf),
     )
     message = 'kumo bot: deleted stack %s ' % _get_stack_name(conf)
-    monitoring.slacker_notification('systemmessages', message, slack_token)
+    monitoring.slack_notification(slack_channel, message, slack_token)
     stackname = _get_stack_name(conf)
     return _poll_stack_events(boto_session, stackname)
 
@@ -598,35 +606,3 @@ def generate_template_file(conf, cloudformation):
         opened_file.write(template_body)
     print('wrote cf-template for %s to disk: %s' % (get_env(), template_file_name))
     return template_file_name
-
-
-def configure(config_file=None):
-    """Create the .gcdt config file in the users home folder.
-
-    :param config_file:
-    """
-    if not config_file:
-        config_file = os.path.expanduser('~') + '/' + '.kumo'
-    stack_name = _get_input()
-    with open(config_file, 'w') as config:
-        config.write('kumo {\n')
-        config.write('slack-token=%s' % stack_name)
-        config.write('\n}')
-
-
-def read_kumo_config(config_file=None):
-    """Read .kumo config file from user home.
-
-    :return: pyhocon configuration, exit_code
-    """
-    if not config_file:
-        config_file = os.path.expanduser('~') + '/' + '.kumo'
-    try:
-        config = ConfigFactory.parse_file(config_file)
-        return config, 0
-    except Exception as e:
-        print(e)
-        print(colored.red('Cannot find file .kumo in your home directory %s' %
-                          config_file))
-        print(colored.red("Please run 'kumo configure'"))
-        return None, 1