+++ gcdt/monitoring.py
@@ -1,37 +1,26 @@
 # -*- coding: utf-8 -*-
-from time import time
-from functools import wraps
-import boto3
-import datetime
-import json
+from __future__ import print_function
+import sys
+from clint.textui import colored
 from slacker import Slacker
 from gcdt.logger import setup_logger
 
-log = setup_logger(logger_name="monitoring")
+log = setup_logger(__name__)
 
 
-# TODO: this needs tests!
-
-def get_cloudwatch_client():
-    return boto3.client("cloudwatch")
-
-
-def get_cloudwatch_events_client():
-    return boto3.client('events')
-
-
-def put_metrics(namespace, metric_data):
-    cloudwatch = get_cloudwatch_client()
+''' currently not used?
+def _put_metrics(namespace, metric_data):
+    cloudwatch = boto3.client("cloudwatch")
     cloudwatch.put_metric_data(Namespace=namespace, MetricData=metric_data)
 
 
-def timing(namespace, metric, value):
+def _timing(namespace, metric, value):
     """Record a timing.
 
-    >>> monitoring.timing("query.response.time", 1234)
+    monitoring.timing("query.response.time", 1234)
     """
     print(metric, 'ms', value)
-    put_metrics(
+    _put_metrics(
         Namespace=namespace,
         MetricData=[
             {'MetricName': metric,
@@ -55,22 +44,23 @@ def timed(namespace, metric):
         finally:
             monitoring.timing('user.query.time', time.time() - start)
     """
-
     def wrapper(func):
         @wraps(func)
         def wrapped(*args, **kwargs):
             start = time()
             result = func(*args, **kwargs)
-            timing(namespace, metric, time() - start)
+            _timing(namespace, metric, time() - start)
             return result
 
         return wrapped
 
     return wrapper
+'''
 
 
-def push_event(namespace, cloudwatch_event):
-    client = get_cloudwatch_events_client()
+''' currently not used?
+def _push_event(namespace, cloudwatch_event):
+    client = boto3.client('events')
     detail = {"description": cloudwatch_event}
     response = client.put_events(
         Entries=[
@@ -98,20 +88,20 @@ def event(namespace, cloudwatch_event):
             pass
 
     """
-
     def wrapper(func):
         @wraps(func)
         def wrapped(*args, **kwargs):
             result = func(*args, **kwargs)
-            push_event(namespace, cloudwatch_event)
+            _push_event(namespace, cloudwatch_event)
             return result
 
         return wrapped
 
     return wrapper
+'''
 
-
-def send_to_slacker(channel, cloudwatch_event, slack_token):
+''' currently not used?
+def send_to_slack(channel, cloudwatch_event, slack_token):
     """A decorator that will push a custom cloudwatch event
 
     @monitoring.event('deploytool', 'deployed xyz')
@@ -119,7 +109,6 @@ def send_to_slacker(channel, cloudwatch_event, slack_token):
         # Do what you need to ...
         pass
     """
-
     if slack_token:
         slack = Slacker(slack_token)
 
@@ -127,19 +116,21 @@ def send_to_slacker(channel, cloudwatch_event, slack_token):
             @wraps(func)
             def wrapped(*args, **kwargs):
                 result = func(*args, **kwargs)
-                slack.chat.post_message('#' + channel, cloudwatch_event)
+                slack.chat.post_message('#%s' % channel, cloudwatch_event)
                 return result
 
             return wrapped
 
         return wrapper
-    else:
-        pass
+'''
 
 
-def slacker_notification(channel, message, slack_token):
-    if slack_token:
-        slack = Slacker(slack_token)
-        slack.chat.post_message('#' + channel, message)
-    else:
-        pass
+# TODO: maybe move to utils?
+def slack_notification(channel, message, slack_token, out=sys.stdout):
+    if channel and slack_token:
+        try:
+            slack = Slacker(slack_token)
+            slack.chat.post_message('#%s' % channel, message)
+        except Exception as e:
+            print(colored.red('We can not use your slack token: %s' % str(e)),
+                  file=out)