+++ tests/test_utils.py
@@ -1,7 +1,11 @@
 # -*- coding: utf-8 -*-
+import os
+from tempfile import NamedTemporaryFile
 from nose.tools import assert_equal
 from StringIO import StringIO
-from gcdt.utils import version, __version__, retries
+from gcdt import utils
+from gcdt.utils import version, __version__, retries, configure, \
+    read_gcdt_user_config
 
 
 def test_version():
@@ -71,3 +75,52 @@ def test_retries_raises_exception():
 
     assert_equal(state['r'], 5)
     assert_equal(state['h'], 4)
+
+
+def test_configure():
+    stackname = 'my_stack'
+
+    def fake_get_input():
+        return stackname
+
+    utils._get_slack_token_from_user = fake_get_input
+
+    tf = NamedTemporaryFile(delete=False)
+    configure(tf.name)
+    assert_equal(open(tf.name).read(), 'gcdt {\nslack-token=%s\n}' % stackname)
+
+    # cleanup the testfile
+    tf.close()
+    os.unlink(tf.name)
+
+
+def test_read_user_config():
+    expected_slack_token = 'my_slack_token'
+    expected_slack_channel = 'my_slack_channel'
+
+    tf = NamedTemporaryFile(delete=False)
+    open(tf.name, 'w').write('gcdt {\nslack-token=%s\nslack-channel=%s\n}' %
+                             (expected_slack_token, expected_slack_channel))
+
+    slack_token, slack_channel = read_gcdt_user_config(tf.name)
+    assert_equal(slack_token, expected_slack_token)
+    assert_equal(slack_channel, expected_slack_channel)
+
+    # cleanup the testfile
+    tf.close()
+    os.unlink(tf.name)
+
+
+def test_read_user_config_comp_mode():
+    expected_slack_token = 'my_slack_token'
+
+    tf = NamedTemporaryFile(delete=False)
+    open(tf.name, 'w').write('kumo {\nslack-token=%s\n}' % expected_slack_token)
+
+    slack_token, slack_channel = read_gcdt_user_config(tf.name, 'kumo')
+    assert_equal(slack_token, expected_slack_token)
+    assert_equal(slack_channel, 'systemmessages')
+
+    # cleanup the testfile
+    tf.close()
+    os.unlink(tf.name)